examples/03-validation/src/main/java/com/intelematics/architecture/typesafe/configuration/Configuration.java
package com.intelematics.architecture.typesafe.configuration;

import org.springframework.boot.context.properties.ConfigurationProperties;
import org.springframework.validation.annotation.Validated;

import javax.validation.constraints.NotBlank;

@ConfigurationProperties("example.configuration")
@Validated
public class Configuration
{
    @NotBlank
    private String value;

    public String getValue()
    {
        return value;
    }

    public void setValue(String value)
    {
        this.value = value;
    }
}
